// src/api-client/models/verification-pass-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface VerificationPassResponseDto
 */
export interface VerificationPassResponseDto {
    /**
     * The recipient.
     * @type {string}
     * @memberof VerificationPassResponseDto
     */
    'recipientId': string;
    /**
     * Public name.
     * @type {string}
     * @memberof VerificationPassResponseDto
     */
    'displayName': string;
    /**
     * Charity registration number, shown as the organization ID.
     * @type {string}
     * @memberof VerificationPassResponseDto
     */
    'registrationCode'?: string | null;
    /**
     * What the QR encodes. Stable for the organization, so it can be printed once.
     * @type {string}
     * @memberof VerificationPassResponseDto
     */
    'passCode': string;
    /**
     * Whether an admin has verified them.
     * @type {boolean}
     * @memberof VerificationPassResponseDto
     */
    'isVerified': boolean;
    /**
     * Initials, for the avatar placeholder.
     * @type {string}
     * @memberof VerificationPassResponseDto
     */
    'initials': string;
}