// src/api-client/models/contact-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Which function this contact covers.
 * @export
 * @enum {string}
 */

export const ContactType = {
    Primary: 'PRIMARY',
    Operations: 'OPERATIONS',
    Logistics: 'LOGISTICS',
    Billing: 'BILLING',
    Driver: 'DRIVER',
    Emergency: 'EMERGENCY'
} as const;

export type ContactType = typeof ContactType[keyof typeof ContactType];