// src/api-client/models/urgency-threshold.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Which surplus urgencies they want to hear about.
 * @export
 * @enum {string}
 */

export const UrgencyThreshold = {
    All: 'ALL',
    CriticalExpiring: 'CRITICAL_EXPIRING',
    CriticalOnly: 'CRITICAL_ONLY'
} as const;

export type UrgencyThreshold = typeof UrgencyThreshold[keyof typeof UrgencyThreshold];